refactor(test): address review feedback on regenerator gating
- Move SkipUnlessUpdatingFixtures and WriteLegAknFixtureIfChanged out of
  DocumentHelpers into a dedicated TestFileUpdateHelpers class.
- Use /// XML doc comments on the public helpers so IntelliSense surfaces them.
- Only save images and log "Regenerated" when the .akn actually changed.
