+++ test/leg/TestEMHtml.cs
@@ -56,7 +56,7 @@ public class TestEMHtml {
 
     [Fact]
     public void RegenerateAllHtml() {
-        DocumentHelpers.SkipUnlessUpdatingFixtures();
+        TestFileUpdateHelpers.SkipUnlessUpdatingFixtures();
         if (!HtmlBuilder.IsAvailable())
             Assert.Skip("HtmlBuilder unavailable - install Oxygen or set OXYGEN_HOME.");
