+++ test/leg/TestIAHtml.cs
@@ -60,7 +60,7 @@ public class TestIAHtml {
 
     [Fact]
     public void RegenerateAllHtml() {
-        DocumentHelpers.SkipUnlessUpdatingFixtures();
+        TestFileUpdateHelpers.SkipUnlessUpdatingFixtures();
         if (!HtmlBuilder.IsAvailable())
             Assert.Skip("HtmlBuilder unavailable - install Oxygen or set OXYGEN_HOME.");
